replace the big decimal with double

+++ src/entity/MonthReport.kt
@@ -1,10 +1,9 @@
 package entity
 
-import java.math.BigDecimal
 
 data class MonthReport(
-    val incomeTotal: BigDecimal,
-    val expenseTotal: BigDecimal,
-    val balance: BigDecimal,
+    val incomeTotal: Double,
+    val expenseTotal: Double,
+    val balance: Double,
     val month: Int,
 )
